Make setup.py executable
Use env to choose python3, and make setup.py executable so that it
can be run in the common "./setup.py" way.  Remove the code that
removed --single-version-externally-managed from arguments; it
can be added again later if there is a specific reason for it.

Update .gitignore to future-proof for pyvenv use likely for use
with ./setup.py

Change-Id: I1a584d12c58e862dc8d905dec328eb7d1aaaaec8