remove explicit default of WORK so USER works
No longer explicitly specify WORK as the default libref. If USER is assigned, it should be used for single level table names. Let the libref be blank and only code table name in SAS code. When creating a SASdata object though with no libref, determine if the table is in USER or WORK and set the attribute of the object so that the libref is known for that object. Need that for the http interface; it requires libref as part of it's path structure.