Removed the hard-coded link to saspyiom.jar
The config file previously contained a hard-coded link to saspyiom.jar at: "C:\\ProgramData\\Anaconda3\\Lib\\site-packages\\saspy\\java\\saspyiom.jar"

This will break if the user installed SASPy in *any way* other than through Anaconda3 with the default install path, and should not be hard-coded like this.  This changes the config file to look for saspyiom in the same directory that the config file is being read from by accessing the __file__ attribute of the config file.