No mutable default arguments (#181)
Hey, that all looks good. That's interesting that the default values are global, not instantiated per invocation. I didn't know that; rookie mistake :). I bet that is an interesting problem to track down if there was a case where they were modified.
Thanks for cleaning this up for me and contributing!
Tom