Add IO test suite. See log for details.
Added tests to validate SAS IO related functionality. Currently, the
only tests written are to check for IO object compliance. That is, the
tests make sure that each IO object has the required methods defined so
that all objects expose a coherent interface. This was particularly
tricky when defining the COM module as some private functions need to be
exposed publicly (`_asubmit`).

Additional tests would be useful.