Improve Unittest File for STATS Products (#519)
* Rename Test Functions So They Run in UnitTest

* factorResults Mislabeled

The factorResultX functions fail unit testing. This is due to them
accidentally calling PROC REG instead of PROC FACTOR.

* Add missing imports to test_sasstat

* Fix Incorrect PROC Call in Test for TTest

* Skipping Seemingly Incorrect Test

* Fix Malformed Test for STRDSET1

* Skip regResult4

This test produces the correct output in Jupyter, but has the wrong
type. This should be looked into more, not sure what's wrong.

* Fix Misleading and Incomplete Assert Messages

* Conditionally Skip Tests involving IPython and Pandas

STAT test modified to conditionally skip certain tests if either IPython
or Pandas are not available. Each package is checked separately.